requirements.txt
ipython
pandas
pyarrow
numpy
pyspark
kafka-python
py4j